add ExpressionReplacer
